fix: add id key and pagination meta object at /posts/search/sources api (#130)
